backend/requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Redis
redis==5.0.1
hiredis==2.2.3

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.1
google-auth==2.25.2
google-auth-oauthlib==1.1.0
twilio==8.10.3

# ML/AI - Using lightweight versions for Docker
# Note: TensorFlow removed for Docker size. Using mock predictions.
# For production with real ML model, install: tensorflow==2.15.0
Pillow==10.1.0
numpy==1.26.2
# opencv-python removed for Docker size (heavy dependency)
google-generativeai==0.3.1

# Utilities
python-dotenv==1.0.0
pydantic==2.5.2
pydantic-settings==2.1.0
email-validator==2.1.0
httpx==0.25.2
